+++ ShapeEngine/Stats/IStat.cs
@@ -1,0 +1,8 @@
+namespace ShapeEngine.Stats;
+
+public interface IStat
+{
+    public void Reset();
+    public bool IsAffected(uint tag);
+    public void Apply(BuffValue buffValue);
+}
